(fix): site admins can delete group; fix permissions when deleting (#293)

+++ plugins/Groups/Controllers/api/v1/group.php
@@ -196,6 +196,12 @@ class group implements Interfaces\Api
             return Factory::response([]);
         }
 
+        if (!Session::isAdmin() && !$group->isOwner($user)) {
+            return Factory::response([
+                'error' => 'You cannot delete this group'
+            ]);
+        }
+
         try {
             $group->delete();
